requirements.txt
numpy<2.0
torch
pyexr
Pillow
tqdm
tensorboard